requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.20.0
anthropic>=0.5.0
openai>=1.0.0
pinecone-client>=2.0.0
pydantic>=1.10.8
tiktoken>=0.3.0
python-jose>=3.2.0
sqlalchemy>=1.4.46
asyncpg>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
tenacity>=8.0.0